+++ src/app/src/components/Filters/CountryNameFilter.jsx
@@ -19,7 +19,11 @@ function CountryNameFilter({
     return (
         <div className="form__field">
             <StyledSelect
-                label="Country Name"
+                label={
+                    <div style={{ display: 'flex' }}>
+                        <p>Country Name</p>
+                    </div>
+                }
                 name={COUNTRIES}
                 options={countryOptions || []}
                 value={countries}