+++ src/app/src/components/MapAndSidebar.jsx
@@ -1,123 +1,24 @@
-import React, { Component, Fragment } from 'react';
-import { Route, Switch } from 'react-router-dom';
+import React, { Fragment } from 'react';
+import { Route } from 'react-router-dom';
+import { Hidden } from '@material-ui/core';
 import Grid from '@material-ui/core/Grid';
 
 import SidebarWithErrorBoundary from './SidebarWithErrorBoundary';
-import FacilitiesMap from './FacilitiesMap';
-import FacilitiesMapErrorMessage from './FacilitiesMapErrorMessage';
-import FeatureFlag from './FeatureFlag';
-import VectorTileFacilitiesMap from './VectorTileFacilitiesMap';
 
 import '../styles/css/Map.css';
+import Map from './Map';
 
-import withQueryStringSync from '../util/withQueryStringSync';
-
-import {
-    mainRoute,
-    facilitiesRoute,
-    facilityDetailsRoute,
-    VECTOR_TILE,
-} from '../util/constants';
-
-class MapAndSidebar extends Component {
-    state = { hasError: false };
-
-    static getDerivedStateFromError() {
-        return { hasError: true };
-    }
-
-    componentDidCatch(error) {
-        if (window.Rollbar) {
-            window.Rollbar.error(error);
-        }
-    }
-
-    render() {
-        const { hasError } = this.state;
-
-        return (
-            <Fragment>
-                <Grid container className="map-sidebar-container">
-                    <Grid item sm={12} md={4} id="panel-container">
-                        <Route component={SidebarWithErrorBoundary} />
-                    </Grid>
-                    <Grid
-                        item
-                        sm={12}
-                        md={8}
-                        style={{ position: 'relative' }}
-                        className="map-container"
-                    >
-                        {!hasError && (
-                            <Switch>
-                                <Route
-                                    exact
-                                    path={facilityDetailsRoute}
-                                    render={() => (
-                                        <FeatureFlag
-                                            flag={VECTOR_TILE}
-                                            alternative={
-                                                <Route
-                                                    component={FacilitiesMap}
-                                                />
-                                            }
-                                        >
-                                            <Route
-                                                component={
-                                                    VectorTileFacilitiesMap
-                                                }
-                                            />
-                                        </FeatureFlag>
-                                    )}
-                                />
-                                <Route
-                                    exact
-                                    path={facilitiesRoute}
-                                    render={() => (
-                                        <FeatureFlag
-                                            flag={VECTOR_TILE}
-                                            alternative={
-                                                <Route
-                                                    component={FacilitiesMap}
-                                                />
-                                            }
-                                        >
-                                            <Route
-                                                component={
-                                                    VectorTileFacilitiesMap
-                                                }
-                                            />
-                                        </FeatureFlag>
-                                    )}
-                                />
-                                <Route
-                                    exact
-                                    path={mainRoute}
-                                    render={() => (
-                                        <FeatureFlag
-                                            flag={VECTOR_TILE}
-                                            alternative={
-                                                <Route
-                                                    component={FacilitiesMap}
-                                                />
-                                            }
-                                        >
-                                            <Route
-                                                component={
-                                                    VectorTileFacilitiesMap
-                                                }
-                                            />
-                                        </FeatureFlag>
-                                    )}
-                                />
-                            </Switch>
-                        )}
-                        {hasError && <FacilitiesMapErrorMessage />}
-                    </Grid>
-                </Grid>
-            </Fragment>
-        );
-    }
+function MapAndSidebar() {
+    return (
+        <Fragment>
+            <Grid container className="map-sidebar-container">
+                <Route component={SidebarWithErrorBoundary} />
+                <Hidden mdDown>
+                    <Map />
+                </Hidden>
+            </Grid>
+        </Fragment>
+    );
 }
 
-export default withQueryStringSync(MapAndSidebar);
+export default MapAndSidebar;