+++ CHANGELOG.md
@@ -68,6 +68,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Rename makecsvs to makedata and add stub SQL mode [#2108](https://github.com/open-apparel-registry/open-apparel-registry/pull/2108)
 - Update facility lists for contributor workflow updates [#2050](https://github.com/open-apparel-registry/open-apparel-registry/issues/2050)
 - Update facility details content [#2126](https://github.com/open-apparel-registry/open-apparel-registry/pull/2126)
+- Search/Filtering/Results UX Updates [#2110](https://github.com/open-apparel-registry/open-apparel-registry/pull/2110)
 
 ### Deprecated